+++ NEWS.md
@@ -1,3 +1,5 @@
+# hubValidations (development)
+
 # hubValidations 0.12.1
 
 ## Bug fixes