+++ NEWS.md
@@ -1,4 +1,4 @@
-# hubValidations (development)
+# hubValidations 0.13.0
 
 ## Enhancements